--- a/scio-core/src/main/scala/com/spotify/scio/coders/KryoAtomicCoder.scala
+++ b/scio-core/src/main/scala/com/spotify/scio/coders/KryoAtomicCoder.scala
@@ -17,9 +17,9 @@
 
 package com.spotify.scio.coders
 
-import java.io.{ByteArrayInputStream, InputStream, OutputStream}
-import java.nio.ByteBuffer
+import java.io.{InputStream, OutputStream}
 
+import com.esotericsoftware.kryo.io.{InputChunked, OutputChunked}
 import com.google.common.io.{ByteStreams, CountingOutputStream}
 import com.google.common.reflect.ClassPath
 import com.google.protobuf.Message
@@ -28,10 +28,9 @@ import com.twitter.chill.algebird.AlgebirdRegistrar
 import com.twitter.chill.protobuf.ProtobufSerializer
 import org.apache.avro.generic.GenericRecord
 import org.apache.avro.specific.SpecificRecordBase
-import org.apache.beam.sdk.coders.Coder.Context
 import org.apache.beam.sdk.coders._
 import org.apache.beam.sdk.io.gcp.bigquery.TableRowJsonCoder
-import org.apache.beam.sdk.util.VarInt
+import org.apache.beam.sdk.util.{EmptyOnDeserializationThreadLocal, VarInt}
 import org.apache.beam.sdk.util.common.ElementByteSizeObserver
 import org.joda.time.{LocalDate, LocalDateTime}
 import org.slf4j.LoggerFactory
@@ -71,60 +70,74 @@ private object KryoRegistrarLoader {
 }
 
 private[scio] class KryoAtomicCoder[T] extends AtomicCoder[T] {
+  private val bufferSize = 64 * 1024
 
   import KryoAtomicCoder.logger
 
-  @transient
-  private lazy val kryo: ThreadLocal[Kryo] = new ThreadLocal[Kryo] {
-    override def initialValue(): Kryo = {
-      val k = KryoSerializer.registered.newKryo()
+  private val kryoState: ThreadLocal[KryoState] =
+    new EmptyOnDeserializationThreadLocal[KryoState] {
+      override def initialValue(): KryoState = {
+        val k = KryoSerializer.registered.newKryo()
 
-      k.forClass(new CoderSerializer(InstantCoder.of()))
-      k.forClass(new CoderSerializer(TableRowJsonCoder.of()))
+        k.forClass(new CoderSerializer(InstantCoder.of()))
+        k.forClass(new CoderSerializer(TableRowJsonCoder.of()))
 
-      // java.lang.Iterable.asScala returns JIterableWrapper which causes problem.
-      // Treat it as standard Iterable instead.
-      k.register(classOf[JIterableWrapper[_]], new JIterableWrapperSerializer())
+        // java.lang.Iterable.asScala returns JIterableWrapper which causes problem.
+        // Treat it as standard Iterable instead.
+        k.register(classOf[JIterableWrapper[_]], new JIterableWrapperSerializer())
 
-      k.forSubclass[SpecificRecordBase](new SpecificAvroSerializer)
-      k.forSubclass[GenericRecord](new GenericAvroSerializer)
-      k.forSubclass[Message](new ProtobufSerializer)
+        k.forSubclass[SpecificRecordBase](new SpecificAvroSerializer)
+        k.forSubclass[GenericRecord](new GenericAvroSerializer)
+        k.forSubclass[Message](new ProtobufSerializer)
 
-      k.forSubclass[LocalDateTime](new JodaLocalDateTimeSerializer)
-      k.forSubclass[LocalDate](new JodaLocalDateSerializer)
+        k.forSubclass[LocalDateTime](new JodaLocalDateTimeSerializer)
+        k.forSubclass[LocalDate](new JodaLocalDateSerializer)
 
-      k.forClass(new KVSerializer)
-      // TODO:
-      // TimestampedValueCoder
+        k.forClass(new KVSerializer)
+        // TODO:
+        // TimestampedValueCoder
 
-      new AlgebirdRegistrar()(k)
-      KryoRegistrarLoader.load(k)
+        new AlgebirdRegistrar()(k)
+        KryoRegistrarLoader.load(k)
 
-      k
+        val input = new InputChunked(bufferSize)
+        val output = new OutputChunked(bufferSize)
+
+        KryoState(k, input, output)
+      }
     }
-  }
 
+  private val header = -1
 
-  override def encode(value: T, outStream: OutputStream): Unit = {
+  override def encode(value: T, os: OutputStream): Unit = {
     if (value == null) {
       throw new CoderException("cannot encode a null value")
     }
-    val os = new BufferedPrefixOutputStream(outStream)
-    val output = new Output(os)
-    kryo.get().writeClassAndObject(output, value)
-    output.flush()
-    os.finish()
+
+    VarInt.encode(header, os)
+
+    val state = kryoState.get()
+
+    val chunked = state.output
+    chunked.setOutputStream(os)
+
+    state.kryo.writeClassAndObject(chunked, value)
+    chunked.endChunks()
+    chunked.flush()
   }
 
-  override def decode(inStream: InputStream): T = {
-    val o = if (VarInt.decodeInt(inStream) == -1) {
-      val is = new BufferedPrefixInputStream(inStream)
-      val obj = kryo.get().readClassAndObject(new Input(is))
-      is.finish()
-      obj
+  override def decode(is: InputStream): T = {
+    val state = kryoState.get()
+
+    val o = if (VarInt.decodeInt(is) == header) {
+      val chunked = state.input
+      chunked.setInputStream(is)
+
+      state.kryo.readClassAndObject(chunked)
     } else {
-      kryo.get().readClassAndObject(new Input(inStream))
+      state.kryo.readClassAndObject(new Input(state.input.getBuffer))
     }
+
     o.asInstanceOf[T]
   }
 
@@ -178,118 +191,20 @@ private[scio] class KryoAtomicCoder[T] extends AtomicCoder[T] {
   private def kryoEncodedElementByteSize(obj: Any): Long = {
     val s = new CountingOutputStream(ByteStreams.nullOutputStream())
     val output = new Output(s)
-    kryo.get().writeClassAndObject(output, obj)
+    kryoState.get().kryo.writeClassAndObject(output, obj)
     output.flush()
     s.getCount + VarInt.getLength(s.getCount)
   }
 
 }
 
+/** Used for sharing Kryo instance and buffers */
+private[scio] final case class KryoState(kryo: Kryo, input: InputChunked, output: OutputChunked)
+
 private[scio] object KryoAtomicCoder {
 
   private val logger = LoggerFactory.getLogger(this.getClass)
 
   def apply[T]: Coder[T] = new KryoAtomicCoder[T]
-
-}
-
-/**
- * Buffered output stream that adds length prefix to each buffer block. Useful for large objects
- * or collections. Based on [[org.apache.beam.sdk.util.BufferedElementCountingOutputStream]].
- */
-private class BufferedPrefixOutputStream(private val os: OutputStream)
-  extends OutputStream {
-
-  private val buffer = ByteBuffer.allocate(64 * 1024)
-  private var finished = false
-
-  VarInt.encode(-1, os)
-
-  override def write(b: Int): Unit = {
-    require(!finished, "Stream has been finished. Can not add any more data.")
-    if (!buffer.hasRemaining) {
-      outputBuffer()
-    }
-    buffer.put(b.toByte)
-  }
-
-  override def write(b: Array[Byte], off: Int, len: Int): Unit = {
-    require(!finished, "Stream has been finished. Can not add any more data.")
-    if (buffer.remaining() >= len) {
-      buffer.put(b, off, len)
-    } else {
-      outputBuffer()
-      if (len < buffer.capacity()) {
-        buffer.put(b, off, len)
-      } else {
-        VarInt.encode(len, os)
-        os.write(b, off, len)
-      }
-    }
-  }
-
-  override def flush(): Unit = if (!finished) {
-    outputBuffer()
-    os.flush()
-  }
-
-  def finish(): Unit = if (!finished) {
-    flush()
-    VarInt.encode(0, os)
-    finished = true
-  }
-
-  private def outputBuffer(): Unit = if (buffer.position() > 0) {
-    VarInt.encode(buffer.position(), os)
-    os.write(buffer.array(), buffer.arrayOffset(), buffer.position())
-    buffer.clear()
-  }
-
-}
-
-/** Counterpart for [[BufferedPrefixOutputStream]]. */
-private class BufferedPrefixInputStream(private val is: InputStream) extends InputStream {
-  inputBuffer()
-
-  private var buffer: Array[Byte] = _
-  private var bais: ByteArrayInputStream = _
-  private var finished = false
-
-  override def read(): Int = {
-    inputBuffer()
-    bais.read()
-  }
-
-  override def read(b: Array[Byte], off: Int, len: Int): Int = {
-    var bytesRead = 0
-    var n = 0
-    do {
-      inputBuffer()
-      n = bais.read(b, off + bytesRead, len - bytesRead)
-      if (n > 0) {
-        bytesRead += n
-      }
-    } while (n > 0)
-    bytesRead
-  }
-
-  private def inputBuffer(): Unit = if (!finished) {
-    if (bais == null || bais.available() <= 0) {
-      val len = VarInt.decodeInt(is)
-      if (len == 0) {
-        finished = true
-      } else {
-        if (buffer == null || buffer.length < len) {
-          buffer = Array.ofDim[Byte](math.max(len, 64 * 1024))
-        }
-        is.read(buffer, 0, len)
-        bais = new ByteArrayInputStream(buffer, 0, len)
-      }
-    }
-  }
-
-  def finish(): Unit = if (!finished) {
-    require(VarInt.decodeInt(is) == 0, "Invalid end of input stream")
-  }
 
 }