Looks like there is performance regression introduced in https://github.com/spotify/scio/pull/478. With this fix, shuffle takes 5-10 times less. Add `scio-jmh` project for micro benchmarks for further improvements.

In one of our pipelines, a shuffle of ~1B records with 16 bytes key takes 4.5 hours instead of 22 hours before.

I propose to handle the case of 400 Mb key separately, cover it with benchmarks, and check if there is a regression in benchmarks for the case of a small key. One of other ways would be to explicitly ask users to specify different coder in case key is huge.

This fix is going to save us approximately 800,000 vCPU hours in an upcoming month, I don't mind backporting if we can ship it as 0.4.0.1.

## Benchmark

```
jmh:run -i 20 -wi 20 -f4 -t1
```

## Before

```
Benchmark                                  Mode  Cnt     Score     Error  Units
KryoAtomicCoderBenchmark.customDecode      avgt   80   286.811 ±   7.392  ns/op
KryoAtomicCoderBenchmark.customEncode      avgt   80   847.654 ±  11.846  ns/op
KryoAtomicCoderBenchmark.customKryoDecode  avgt   80  6184.079 ± 100.342  ns/op
KryoAtomicCoderBenchmark.customKryoEncode  avgt   80  6490.695 ± 196.628  ns/op
KryoAtomicCoderBenchmark.javaDecode        avgt   80  7787.809 ±  58.954  ns/op
KryoAtomicCoderBenchmark.javaEncode        avgt   80  6309.093 ± 105.909  ns/op
KryoAtomicCoderBenchmark.kryoDecode        avgt   80  5851.015 ± 126.595  ns/op
KryoAtomicCoderBenchmark.kryoEncode        avgt   80  6510.682 ± 155.126  ns/op
```

## After

```
Benchmark                                  Mode  Cnt     Score    Error  Units
KryoAtomicCoderBenchmark.customDecode      avgt   80   259.765 ±  1.359  ns/op
KryoAtomicCoderBenchmark.customEncode      avgt   80   813.831 ±  5.095  ns/op
KryoAtomicCoderBenchmark.customKryoDecode  avgt   80   224.362 ±  8.869  ns/op
KryoAtomicCoderBenchmark.customKryoEncode  avgt   80   998.436 ±  8.312  ns/op
KryoAtomicCoderBenchmark.javaDecode        avgt   80  7745.240 ± 54.849  ns/op
KryoAtomicCoderBenchmark.javaEncode        avgt   80  5854.949 ± 51.330  ns/op
KryoAtomicCoderBenchmark.kryoDecode        avgt   80   318.231 ±  2.674  ns/op
KryoAtomicCoderBenchmark.kryoEncode        avgt   80  1055.268 ±  6.411  ns/op
```