@kanterov 2017-09-02
According to benchmark, it looks like there is a big room for improvement for encoding