@kanterov 2017-09-02
Tests are failing because code assumes that `isWholeStream=true`, but this property has been deprecated, going to check other ways to improve this.