@kanterov 2017-09-02
Fixed compatibility, new benchmark

```
Benchmark                                  Mode  Cnt     Score     Error  Units
KryoAtomicCoderBenchmark.customDecode      avgt   80   254.170 ±   2.202  ns/op
KryoAtomicCoderBenchmark.customEncode      avgt   80   789.309 ±  12.876  ns/op
KryoAtomicCoderBenchmark.customKryoDecode  avgt   80   416.311 ±   3.471  ns/op
KryoAtomicCoderBenchmark.customKryoEncode  avgt   80  1145.902 ±   8.498  ns/op
KryoAtomicCoderBenchmark.javaDecode        avgt   80  7769.387 ± 106.154  ns/op
KryoAtomicCoderBenchmark.javaEncode        avgt   80  6197.226 ±  67.831  ns/op
KryoAtomicCoderBenchmark.kryoDecode        avgt   80   545.693 ±   3.605  ns/op
KryoAtomicCoderBenchmark.kryoEncode        avgt   80  1280.221 ±   8.888  ns/op
```