@kanterov 2017-09-02
Got more performance by reusing Kryo buffers

```
Benchmark                                  Mode  Cnt     Score    Error  Units
KryoAtomicCoderBenchmark.customKryoDecode  avgt   80   214.340 ±  9.182  ns/op
KryoAtomicCoderBenchmark.customKryoEncode  avgt   80  1009.785 ± 12.134  ns/op
```