@nevillelyh 2017-09-07
We reverted the change that causes `IndexOutOfBoundsException` and broke this change. Can you rebase please?