@kanterov 2017-09-17
@ravwojdyla Sorry it took me awhile to address review comments because we backported this fix internally, and it became a lower priority for us.

I called project `scio-jmh` because it's an important detail. JMH does code generation to avoid certain JIT optimizations, that affects the whole project and requires specific build setup, that's why such tests should be in a dedicated project, and shouldn't be mixed with other macro benchmarks from #830. As an alternative, we can name it as `scio-bench-jmh` to make it more clear that it does benchmarks.

I still don't fully understand how IOOB problem can happen, but I rebased the code according to master. I guess it shouldn't affect performance a lot because there should be only one instance of coder per executor.

I'm running benchmarks to update numbers.