@kanterov 2017-09-17
@ravwojdyla updated numbers, please take a look again