@ravwojdyla 2017-09-18
@kanterov:
 * `scio-bench-jmh` +1
 * please fix tests
 * `IOOB` is puzzling, in theory it should not happen (and i have validated that at the time of the exception, none of the threads share kryo instances), might be a bug in kryo, there were already similar bugs reported and fixed in kryo. Anyway, I would prefer having kryo in the class for now, until we figure out where exactly is the problem. Also related to this, have validated that if we set references to false, the bug is gone (but obviously that is not a great option).