@kanterov 2017-10-04
@ravwojdyla @nevillelyh there are `scio-tensorflow` tests failing, but this has nothing related to this PR, because they randomly fail on master too, I was able to reproduce this on my machine after several tries.