@kanterov 2017-10-05
Seems like a known issue, https://github.com/spotify/scio/issues/815